Miscellaneous adjustments for warnings and code quality
This commit contains a grab bag of minor changes not worth committing
individually:

* Replaced raw Latin-1 characters with octal escapes to avoid source-file
  encoding issues

* Undefined some convenience macros after they've served their purpose

* Rewrote SEdge::From() to avoid confusing less-capable C++ compilers

* Have oops() print a newline at the end of its message

* Removed "static" keyword from the Bernstein() function definition, as it
  has a non-static prototype in srf/surface.h

* Added casts (and changed a variable type) to quell warnings about integer
  size and signedness

* Simplified an expression with our handy arraylen() macro

+++ style.cpp
@@ -128,7 +128,7 @@ void Style::FreezeDefaultStyles(void) {
 
 DWORD Style::CreateCustomStyle(void) {
     SS.UndoRemember();
-    DWORD vs = max(Style::FIRST_CUSTOM, SK.style.MaximumId() + 1);
+    DWORD vs = max((DWORD)Style::FIRST_CUSTOM, SK.style.MaximumId() + 1);
     hStyle hs = { vs };
     (void)Style::Get(hs);
     return hs.v;
